Made the change committed by Dom in [2050] also for renameInode() and advancedFind(). Removed some leftover imports from a previous update.

+++ src/plugins/org.drftpd.commands.nuke/src/org/drftpd/commands/nuke/NukeBeans.java
@@ -145,7 +145,7 @@ public class NukeBeans {
 		for (Map.Entry<String,Long> entry : nd.getNukees().entrySet()) {
 			String user = entry.getKey();
 			Long l = entry.getValue();
-			list.add(new NukedUser(user, l.longValue()));
+			list.add(new NukedUser(user, l));
 		}
 		return list;
 	}
